cart persisted

+++ Client/src/components/pages/Home/Products.jsx
@@ -3,11 +3,12 @@ import { CiHeart } from "react-icons/ci";
 import { useDispatch, useSelector} from "react-redux";
 import Pagination from "../../common/Pagination";
 import useDebounceHook from "../../../utils/custom-hooks/useDebounce";
-import { addProductInCart, addToCart, removeFromCart } from "../../../redux/actions/productActions";
+import { addProductInCart, addToCart, removeFromCart } from "../../../redux/actions/cartActions";
 import Sorting from "../../common/Sorting"; 
 import Product from "../../common/Product";
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import { Link, Navigate } from "react-router-dom";
 
 const Products = ({ productData, isAddToCart }) => {
     const user = useSelector((state)=> state.role.user)
@@ -47,20 +48,28 @@ const Products = ({ productData, isAddToCart }) => {
     const handleSortingChange = order => {
         setSortOrder(order);
     };
-    
+
+    const auth = JSON.parse(localStorage.getItem('role')) || false;
+    const isAuth = auth.isAuth;
     
 
     const handleClick = product => {
-        if (isAddToCart) {
+        if (isAuth){
+            if (isAddToCart) {
             
-            dispatch(addProductInCart(product))
-          
-        } else {
-            dispatch(removeFromCart(product.id))
-            toast.success("Removed from the cart!", {
-                position: 'top-right',
-              });
+                dispatch(addProductInCart(product))
+              
+            } else {
+                dispatch(removeFromCart(product.id))
+                toast.success("Removed from the cart!", {
+                    position: 'top-right',
+                  });
+            }
+        }else{
+            
+            toast.warning('Please Login!!')
         }
+       
     };
